extract data collection sections
