+++ app/components/Views/Settings/SecuritySettings/SecuritySettings.tsx
@@ -7,7 +7,6 @@ import {
   View,
   ActivityIndicator,
   Keyboard,
-  InteractionManager,
   Platform,
   Linking,
 } from 'react-native';
@@ -19,11 +18,7 @@ import { clearHistory } from '../../../../actions/browser';
 import Logger from '../../../../util/Logger';
 import { getNavigationOptionsTitle } from '../../../UI/Navbar';
 import { setLockTime } from '../../../../actions/settings';
-import { setDataCollectionForMarketing } from '../../../../actions/security';
-import {
-  HOW_TO_MANAGE_METRAMETRICS_SETTINGS,
-  SIMULATION_DETALS_ARTICLE_URL,
-} from '../../../../constants/urls';
+import { SIMULATION_DETALS_ARTICLE_URL } from '../../../../constants/urls';
 import { strings } from '../../../../../locales/i18n';
 import { passwordSet } from '../../../../actions/user';
 import Engine from '../../../../core/Engine';
@@ -95,8 +90,6 @@ import {
   HASH_STRING,
   HASH_TO_TEST,
   IPFS_GATEWAY_SECTION,
-  META_METRICS_DATA_MARKETING_SECTION,
-  META_METRICS_SECTION,
   NFT_AUTO_DETECT_MODE_SECTION,
   NFT_DISPLAY_MEDIA_MODE_SECTION,
   PASSCODE_CHOICE_STRING,
@@ -114,9 +107,6 @@ import Networks, {
 import images from 'images/image-icons';
 import { ETHERSCAN_SUPPORTED_NETWORKS } from '@metamask/transaction-controller/dist/constants';
 import { SecurityPrivacyViewSelectorsIDs } from '../../../../../e2e/selectors/Settings/SecurityAndPrivacy/SecurityPrivacyView.selectors';
-import generateDeviceAnalyticsMetaData, {
-  UserSettingsAnalyticsMetaData as generateUserSettingsAnalyticsMetaData,
-} from '../../../../util/metrics';
 import Text, {
   TextVariant,
   TextColor,
@@ -132,6 +122,7 @@ import BasicFunctionalityComponent from '../../../UI/BasicFunctionality/BasicFun
 import Routes from '../../../../constants/navigation/Routes';
 import { MetaMetrics } from '../../../../core/Analytics';
 import { isTransactionSimulationsFeatureEnabled } from '../../../../util/transaction-controller';
+import MetaMetricsAndDataCollectionSection from './Sections/MetaMetricsAndDataCollectionSection/MetaMetricsAndDataCollectionSection';
 
 const Heading: React.FC<HeadingProps> = ({ children, first }) => {
   const { colors } = useTheme();
@@ -146,7 +137,7 @@ const Heading: React.FC<HeadingProps> = ({ children, first }) => {
 };
 
 const Settings: React.FC = () => {
-  const { trackEvent, isEnabled, enable, addTraitsToUser } = useMetrics();
+  const { trackEvent, isEnabled } = useMetrics();
   const theme = useTheme();
   const { colors } = theme;
   const styles = createStyles(colors);
@@ -193,11 +184,6 @@ const Settings: React.FC = () => {
   const ipfsGateway = useSelector(selectIpfsGateway);
   const isIpfsGatewayEnabled = useSelector(selectIsIpfsGatewayEnabled);
   const myNetworks = ETHERSCAN_SUPPORTED_NETWORKS as EtherscanNetworksType;
-
-  const isDataCollectionForMarketingEnabled = useSelector(
-    (state: any) => state.security.dataCollectionForMarketing,
-  );
-
   const isMainnet = type === MAINNET;
 
   const updateNavBar = useCallback(() => {
@@ -463,138 +449,6 @@ const Settings: React.FC = () => {
           isDisabled={browserHistory.length === 0}
         />
       </View>
-    </View>
-  );
-
-  const toggleMetricsOptIn = async (metricsEnabled: boolean) => {
-    if (metricsEnabled) {
-      const consolidatedTraits = {
-        ...generateDeviceAnalyticsMetaData(),
-        ...generateUserSettingsAnalyticsMetaData(),
-        is_metrics_opted_in: true,
-      };
-      await enable();
-      setAnalyticsEnabled(true);
-
-      InteractionManager.runAfterInteractions(async () => {
-        await addTraitsToUser(consolidatedTraits);
-        trackEvent(MetaMetricsEvents.ANALYTICS_PREFERENCE_SELECTED, {
-          is_metrics_opted_in: true,
-          updated_after_onboarding: true,
-        });
-      });
-    } else {
-      await enable(false);
-      setAnalyticsEnabled(false);
-      if (isDataCollectionForMarketingEnabled) {
-        dispatch(setDataCollectionForMarketing(false));
-      }
-      Alert.alert(
-        strings('app_settings.metametrics_opt_out'),
-        strings('app_settings.metametrics_restart_required'),
-      );
-    }
-  };
-
-  const addMarketingConsentToTraits = (marketingOptIn: boolean) => {
-    InteractionManager.runAfterInteractions(async () => {
-      const traits = {
-        has_marketing_consent: marketingOptIn,
-      };
-      await addTraitsToUser(traits);
-      trackEvent(MetaMetricsEvents.ANALYTICS_PREFERENCE_SELECTED, {
-        ...traits,
-        location: 'settings',
-      });
-    });
-  };
-
-  const toggleDataCollectionForMarketing = async (value: boolean) => {
-    if (value) {
-      if (!analyticsEnabled) {
-        await toggleMetricsOptIn(true);
-      }
-      addMarketingConsentToTraits(value);
-    } else {
-      navigation.navigate(Routes.MODAL.ROOT_MODAL_FLOW, {
-        screen: Routes.SHEET.DATA_COLLECTION,
-      });
-
-      if (analyticsEnabled) {
-        addMarketingConsentToTraits(value);
-      }
-    }
-    dispatch(setDataCollectionForMarketing(value));
-  };
-
-  const renderMetaMetricsSection = () => (
-    <View style={styles.halfSetting} testID={META_METRICS_SECTION}>
-      <View style={styles.titleContainer}>
-        <Text variant={TextVariant.BodyLGMedium} style={styles.title}>
-          {strings('app_settings.metametrics_title')}
-        </Text>
-        <View style={styles.switchElement}>
-          <Switch
-            value={analyticsEnabled}
-            onValueChange={toggleMetricsOptIn}
-            trackColor={{
-              true: colors.primary.default,
-              false: colors.border.muted,
-            }}
-            thumbColor={theme.brandColors.white000}
-            style={styles.switch}
-            ios_backgroundColor={colors.border.muted}
-            testID={SecurityPrivacyViewSelectorsIDs.METAMETRICS_SWITCH}
-          />
-        </View>
-      </View>
-      <Text
-        variant={TextVariant.BodyMD}
-        color={TextColor.Alternative}
-        style={styles.desc}
-      >
-        {strings('app_settings.metametrics_description')}{' '}
-        <Button
-          variant={ButtonVariants.Link}
-          size={ButtonSize.Auto}
-          onPress={() => Linking.openURL(HOW_TO_MANAGE_METRAMETRICS_SETTINGS)}
-          label={strings('app_settings.learn_more')}
-        />
-      </Text>
-    </View>
-  );
-
-  const renderDataCollectionSection = () => (
-    <View
-      style={styles.halfSetting}
-      testID={META_METRICS_DATA_MARKETING_SECTION}
-    >
-      <View style={styles.titleContainer}>
-        <Text variant={TextVariant.BodyLGMedium} style={styles.title}>
-          {strings('app_settings.data_collection_title')}
-        </Text>
-        <View style={styles.switchElement}>
-          <Switch
-            value={isDataCollectionForMarketingEnabled}
-            onValueChange={toggleDataCollectionForMarketing}
-            trackColor={{
-              true: colors.primary.default,
-              false: colors.border.muted,
-            }}
-            thumbColor={theme.brandColors.white000}
-            style={styles.switch}
-            ios_backgroundColor={colors.border.muted}
-            testID={SecurityPrivacyViewSelectorsIDs.DATA_COLLECTION_SWITCH}
-          />
-        </View>
-      </View>
-      <Text
-        variant={TextVariant.BodyMD}
-        color={TextColor.Alternative}
-        style={styles.desc}
-      >
-        {strings('app_settings.data_collection_description')}
-      </Text>
     </View>
   );
 
@@ -1209,8 +1063,7 @@ const Settings: React.FC = () => {
         >
           {strings('app_settings.analytics_subheading')}
         </Text>
-        {renderMetaMetricsSection()}
-        {renderDataCollectionSection()}
+        <MetaMetricsAndDataCollectionSection />
         <DeleteMetaMetricsData metricsOptin={analyticsEnabled} />
         <DeleteWalletData />
         {renderHint()}
